BlueskyFirehoseReader/requirements.txt
orjson
websocket-client